+++ tests/conftest.py
@@ -1,5 +1,0 @@
-import os
-
-
-def pytest_generate_tests(metafunc):
-    os.environ["ALLOWED_ISS"] = "test"